Provide publisher options for message properties and size variation.
To create tests that more closely mimic actual use cases, the -v
(variation) option specifies the variation in size from the -s
(size) value specified.  The variations are chosen randomly from a
uniform distribution.  If the specified variation is larger than or
equal to the specified size, the variation is silently set to 0.

To more closely mimic actual use cases, the -P (set message properties)
option sets the value of some message properties.  Specifically,

  * app_id is the name of the Publisher class
  * content type is application/octet-stream
  * delivery_mode is 1 (default) or 2 (if persistent flag is set)
  * message_id is the test sequence number of the message
  * priority is 0
  * timestamp is the instant at which the message was constructed

Although these properties may be a small number of bytes compared to
large message payloads, they add to the number of items allocated on
the erlang heap and thus affect garbage collection.