Provide a way to load message payload from disk
A list of comma-separated files can be provided with the -B option.
The content-type of the files can also be specified with the -T option.
If several files are provided, they're used in a round-robin fashion.

Example of usage:
bin/runjava com.rabbitmq.perf.PerfTest -y 0 -u perf.test \
                                       -B file1.json,file2.json -T application/json

This commit makes it easier to test real consumers with some samples of
real messages body.

Fixes #23

[#147758919]