Use default exchange by default
The default exchange is now the default if possible
(e.g. when the routing key isn't specified and
there's only one queue to publish too). For those cases,
the routing key is the queue name and there's no
need to create an exchange (a "direct"-named exchange
was created previously).

Fixes #45