Calculate latency based on header
The latency to receive a message was based on a value in the message payload.
This latency is then used to conditionally display the metrics on received messages.

This worked well as long as the payload contained the time when the message
was sent (the default) but it doesn't work anymore when specifying a given payload.

Feature #23 introduced content based on a file, so the latency wasn't correctly
calculated.

Latency is now calculated based on an header value, which is always sent,
whatever the MessageBodySource is used.

Fixes #50