Spreads producers and consumers across multiple queues based on a queue pattern
This allows creating a sequence of queues and balancing the producers
and the consumers across those queues.

E.g. --queue-pattern 'perf-test-${i}' --from 1 --to 100 --producers 150 --consumers 300

This will run 2 producers per perf-test-[1..50] queue, and 1 producer
per perf-test-[51..100] queue, for a total of 150 producers.
There will be 3 consumers per perf-test-[1..100] queue, for a total of 300 consumers.

This commit contains also a bit of Java 8 refactoring and an upgrade to JUnit 5.

[#153988697,#153991193,#151090466]

Fixes #64