Rationalize threads usage
This commit rationalizes the number of threads used by
a PerfTest instance. This should result in much fewer
threads used for common scenarios and should help to
run tests that simulate high load (many producers and/or consumers).

Previous behavior:

* One thread to run each consumer and producer
* Nb cores * 2 threads created for each consumer and producer
* One thread to send heartbeats for each consumer and producer

New behavior:

* Threading is now handle with ExecutorServices
* PerfTest threads are prefixed with "perf-test-"
* There's only one thread to start all the consumers, as
the PerfTest Consumer just register a Java Client Consumer
and wait.
* Each consumer connection now uses an appropriate number
of threads for its Java Client consumers, e.g. 1 thread for each
channel, and no more than nb cores * 2 if number of channels
is > nb cores * 2. This should behave as previously, but
with fewer threads if the number of channels is small.
* The --heartbeat-sender-threads parameter allows to specify
the number of threads for all the producers and the consumers.
The default is nb producers + nb consumers (to stick to default
behavior), and it can be lowered down significantly manually, as
a few threads should be enough, especially for a performance tool.
* Producers don't allocate any threads for their consumers, as they
don't have any.

[#154174064]

Fixes #70