Add support for exclusive queues with -E flag
This is especially useful with the queue sequence options, but
only if the number of queues and of consumers are the same. If more
queues than consumers are specified, extra queues won't be created.
If more consumers than queues are specified, extra consumers won't
be able to consume from queues, as they don't use the same connection.

[#155934728]
Fixes #76