Support low rates suitable for IoT workloads
This commit adds the --publishing-interval (in seconds) to be able
to specify the number of seconds between 2 published messages for each
producer. It works internally the same way as the publishing rate
feature (it's easier to specify an interval in seconds than a rate
less than 1).
The commit also adds the --producer-random-start-delay (or -prsd) option to
randomly delay the first publishing of each producer. This avoids
a publishing storm when using --publisher-interval.

The following command will create 100 queues/producers/consumers, each
producer will randomly start between 1 and 10 seconds and will then publish a message
every 15 seconds.

-x 100 -y 100 --queue-pattern 'perf-test-%d' --queue-pattern-from 1 --queue-pattern-to 100 \
  --publishing-interval 15 --producer-random-start-delay 10

[#155122722]

References #80