Use scheduling for low rates
With this commit, PerfTest uses a ScheduledExecutorService for producers
as soon as the --publishing-interval option is used. This way a producer
doesn't use a thread of its own, but shares threads with other
producers. This makes the number of threads used much lower when
simulating slow publishers (e.g. a few threads can handle 1000s of
producers).

The original behavior is maintained if --publishing-interval isn't used.

[#155518891,#155122722]

Fixes #80