Support NIO
This commit adds the following options:
 * --nio-threads (-niot): number of NIO threads to use
 * --nio-thread-pool (-niotp): maximum size of the executor service to use for NIO threads.

If --nio-thread-pool isn't specified, PerfTest uses --nio-threads + 2
(which should be OK for most cases), as the thread pool is used to
provide NIO threads, but also threads to close connections.

NIO are appropriate in PerfTest for IoT loads, e.g. with a large number
of producers and consumers with low publishing rate (1 message x seconds
for each producer with the --publishing-interval option). This way
PerfTest can simulate 1000s of clients without using much resource. NIO
in PerfTest are appropriate high throughput loads.

[#154553000]

Fixes #83