Remove com.rabbitmq.examples
Those examples were moved into this repo when it was extracted
from the RabbitMQ Java client one. Most of them are not related to
PerfTest and are many years old. Arguably they are not necessary
today and at best belong in an archive repository of their own.

Per discussion with @acogoluegnes.