Make producer and consumer state thread-safe
This is mainly to deal with rate limit. Rate-limited producer state
doesn't need to be thread-safe, as the producer always executes in the same
thread (this is when publishing more than 1 message per second). For
publishers publishing every x second (with the --publishing-interval
option), the publishing is scheduled in different threads, so the state
must be thread-safe. For consumers, state must be thread-safe, as
message dispatching isn't guaranteed to be in the same thread.