Configure log explicitly with custom configuration file
Jetty (used in Prometheus support) is verbose when starting, so the
default Logback configuration isn't used anymore. As PerfTest cannot
embed a logback.xml file at the root of the classpath, a custom file is
provided and used explicitly. This can be overriden by using the
logback.configurationFile system property.

[#158209647]

References #42