Fix GPG_KEYNAME extraction in macOS
grep (BSD grep) 2.5.1-FreeBSD, the variant found on macOS 10.14
(Mojave), does not have the -P flag. awk to the rescue! awk version
20070501 works just fine.

Notice the lazy evaluation, where the variable will be expanded
just-in-time, when it gets referenced in the make target. This speeds up
execution of all make targets since this gets invoked only when the
variable is used.

The only downside is that this will be invoked multiple times if the
variable is referenced multiple times.