Define phony targets in situ
Having a list of phonies at the end of the Makefile makes it easy to
miss adding a new phony when a target gets defined. Also, when the list
of targets grows, it's more difficult to manage phonies. We are
defining a target as phony just above the target definition - this keeps
things close.