Don't force IPv4 in native binary
The native binary would not start on some environments
(where IPv6 was disabled, but most likely not in a proper way). Forcing
IPv4 would fix the issue, but this would make the binary incompatible
with IPv6.

This seems to have been fixed in SubstrateVM 19.3 (tested successfully
on an environment that would trigger the issue).

References #208

Fixes #217