Allows decimal values for publishing interval
This commit allows values for --publishing-interval to go
below 1, e.g. 0.5 to publish 2 messages per second. The
lower bound is 0.1, to avoid overloading the scheduling.

This allows scheduled publishers to publish up to 10 messages / second,
which can be useful for some scenarios. These scenarios do not have
to be pure IoT scenarios (--publishing-interval was introduced
initially for IoT scenarios), but rather scenarios with a number
of connections and ingress load in mind.

This commit also changes the calculation of the number of scheduling
threads. It assumes a thread is enough to deal with 100 operations
a second (e.g. 100 publishers publishing at 1 message / second), which
is rather a conservative default.

The number of threads can still be changed with the --producer-scheduler-threads
option.

Fixes #287