Re-create binding even for pre-declared queue
Bindings for "non-recorded queues" (that is queues not created
by the run) where not re-created on connection recovery. This
could a problem for non-durable exchanges in case of broker
restart: the exchange would be re-created but not the binding,
so publishers would end up publishing to exchanges without bindings.
So published messages would even be confirmed.

This commit changes this behavior and re-creates the bindings.
The operation may fail (in case the queue is not longer there
for some reason), but this should not stop the recovery.

Fixes #290