Schedule re-subscription if necessary after node restart
Necessary = queue does not exist if classic queue, durable,
and in a cluster setup. In such a case, the queue "does
not exist" as long as the node is down and becomes available
again when the node restarts. Consumers on other nodes
are then scheduled to subscribe for some time.

Fixes #330