Add topology recovery in cancel handler
The consumer cancel handler attempts to re-subscribe when it receives
the cancellation frame. This works for cases when a CQ hosting node is
down, and eventually comes up; or when a CQ is down for some reason, and
later on recovers.

That logic, however, does not account for a topology change e.g. a queue
deletion. A queue deletion also sends a consumer cancel frame. In this
situation, we have to do a partial topology recovery. In this case, we
have to re-declare the queue and its bindings.

The new logic should not impact the previous re-subscription mechanism,
since a channel.queue-declare on a CQ that is down, or inaccessible,
simply returns a channel error, which is captured, and lets the resub
mechanism continue.

Signed-off-by: Aitor Perez Cedres <acedres@vmware.com>