Recover queue and bindings in consumer cancellation
Make sure to use the already existing sequence to recover a queue,
which takes into account server-named queues and their name changing.

Try to recover the queue if it does not exist, even if it's possibly
on a node that is down (and so cannot be re-created).

Add tests.

Fixes #378