Use millisecond granularity for publisher ramp-up period
It was seconds, which works for very long ramp-up period
(e.g. 60 seconds or more) and a small number of publishers
(few 100s), but not short ramp-up period and large numbers
of publishers (start 5000 publishers in 15 seconds). In such
cases, the workload was not random enough like in real-life
and metrics like publisher confirm latency could be much
higher than expected.

References #394