Use atomic utilities instead of synchronization in stats
All the counters are converted to AtomicLong instead of
regular, non-volatile properties updated and read in synchronized
blocks. This avoids a lot of locking.

The Stats class has also been refactored to use nano time
instead of the current time, which is more appropriate and
more accurate to measure intervals.