Bump wrangler
Bumping wrangler to, at least, 3.1.1. The caret ^ in the version will
make sure we keep in the latest version of 3.x major version. There are
no special considerations to upgrade to v3, according to:

https://developers.cloudflare.com/workers/wrangler/migration/update-v2-to-v3/

There is one deprecation that affects us. Wrangler publish is deprecated
in favour of wrangler deploy. Both take same arguments:

https://developers.cloudflare.com/workers/wrangler/deprecations/#publish

Signed-off-by: Aitor Pérez Cedres <acedres@vmware.com>