Disable Datadog metrics test on Semeru
The test does not work on Semeru, maybe a problem
with the way Micrometer send requests to the target URL
(the test would work while debugging, so maybe a race
condition somewhere).

This is a rarely used feature on a JVM less popular
than other, so it's OK to disable the test.