Server-named exclusive queues should not re-use connections
PerfTest re-uses exclusive queue connections when there are more
consumers than queues. This makes sense when queue names are provided by
the user, but cannot apply when no queues names are provided and
server-named queues are assumed. For the latter we assume queue count =
consumer count.

This rule was not enforced before this commit, so all consumers would
share the same connection when using server-named exclusive queues
(because PerfTest was using the queue name argument as the key for an
exclusive queue, and the argument was empty for server-named queues).